add picker module
